feat: add search page
